Moved GigyaJSService to auth module
