+++ projects/gigya/public_api.ts
@@ -3,6 +3,5 @@
  */
 export * from './src/gigya.module';
 export * from './src/cms-components/gigya-raas/index';
-export * from './src/auth/facade/gigya-auth.service';
+export * from './src/auth/facade/index';
 export * from './src/auth/store/actions/index';
-export * from './src/cms-components/gigya-js/facade/gigya-js.service';
