+++ projects/gigya/src/auth/facade/index.ts
@@ -1,0 +1,2 @@
+export * from './gigya-auth.service';
+export * from './gigya-js.service';
