+++ projects/gigya/src/cms-components/gigya-raas/gigya-raas.component.ts
@@ -11,7 +11,7 @@ import { Observable } from 'rxjs';
 import { distinctUntilChanged, take, tap } from 'rxjs/operators';
 import { GigyaConfig } from '../../config';
 import { GigyaRaasComponentData } from '../cms.model';
-import { GigyaJsService } from '../gigya-js/facade/gigya-js.service';
+import { GigyaJsService } from '../../auth/facade/gigya-js.service';
 
 @Component({
   selector: 'cx-gigya-raas',
